Creating StringUtils for separating out the string utility functions from action binder class
